pyproject.toml
[project]
name = "snowflake-admin-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.0",
    "pymongo>=4.13.0",
    "snowflake-connector-python>=3.14.0",
]
